cli/commands/build-cache: separate out cache and options handling
